refactor: add rec component
