+++ src/components/recommendations/recommendations.module.scss
@@ -1,0 +1,3 @@
+.recs {
+	padding: var(--side-padding);
+}
