+++ src/pages/homePage/homePage.module.scss
@@ -11,10 +11,6 @@
 	gap: 5.7rem;
 }
 
-.recs {
-	padding: var(--side-padding);
-}
-
 .films {
 	gap: 1.875rem;
 	padding: var(--side-padding);
